+++ R/rio.R
@@ -25,5 +25,5 @@
 #' 
 #' @references 
 #'   \href{https://github.com/Stan125/GREA}{GREA} provides an RStudio add-in to import data using rio.
-#' @seealso \code{\link{import}}, \code{\link{export}}, \code{\link{convert}}, \code{\link{install_formats}}
+#' @seealso \code{\link{import}}, \code{\link{import_list}}, \code{\link{export}}, \code{\link{convert}}, \code{\link{install_formats}}
 NULL